Refactor: Move models to separate files
- Book
- Chapter
