+++ Audiobook Player/Models/Chapter.swift
@@ -1,0 +1,16 @@
+//
+//  Chapter.swift
+//  Audiobook Player
+//
+//  Created by Florian Pichler on 09.04.18.
+//  Copyright © 2018 Tortuga Power. All rights reserved.
+//
+
+import UIKit
+
+struct Chapter {
+    var title: String
+    var start: Double
+    var duration: TimeInterval
+    var index: Int
+}
